format: All files

+++ script/deploy/SimpleAccount.s.sol
@@ -7,6 +7,7 @@ import "../ScriptUintHelper.sol";
 
 contract DeploySimpleAccountFactory is Script {
     using ScriptUintHelper for uint256;
+
     string _network;
 
     function setUp() public {
@@ -56,7 +57,7 @@ contract DeploySimpleAccountFactory is Script {
         address expectedAddress = factory.getAddress(signer, salt);
         assert(address(account) == expectedAddress);
         console2.log("New simpleAccount address:", expectedAddress);
-        uint nonce = account.getNonce();
+        uint256 nonce = account.getNonce();
         console2.log("Account nonce", nonce);
 
         vm.stopBroadcast(); // End the broadcast session
